Refactor the integration tests setup (kubernetes-sigs/gateway-api-inference-extension#506)
